# [wip] experiment(http): introduce `requestEvent.sendFile` fast path