A new method to handle sending files to the socket. In the future, this could _also_ use the `sendfile(2)` syscall to avoid file copy in the userspace.

`main`:
```
Running 10s test @ http://localhost:8080/README.md
  8 threads and 20 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   759.55us  121.38us   4.67ms   83.20%
    Req/Sec     2.64k    86.82     2.76k    92.20%
  212156 requests in 10.10s, 558.63MB read
Requests/sec:  21005.52
Transfer/sec:     55.31MB
```

This PR (`sendFile`):
```
Running 10s test @ http://localhost:8080/README.md
  8 threads and 20 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   470.52us  126.24us   3.43ms   77.33%
    Req/Sec     4.27k   199.36     4.68k    80.57%
  343083 requests in 10.10s, 0.88GB read
Requests/sec:  33968.47
Transfer/sec:     89.44MB
```

Node:
```
Running 10s test @ http://localhost:8081/README.md
  8 threads and 20 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   610.48us  782.74us  29.18ms   99.41%
    Req/Sec     3.47k   326.92     3.64k    92.08%
  279390 requests in 10.10s, 748.21MB read
Requests/sec:  27662.49
Transfer/sec:     74.08MB
```